Misc-cleanup (#162)
* misc files

* more clean up

+++ evidence/queries/ncaaf/most_recent_games.sql
@@ -3,7 +3,7 @@ SELECT
     RL.visiting_team,
     '@' as " ",
     RL.home_team,
-    RL.home_team_score || ' - ' || RL.visiting_team_score as score,
+    RL.home_team_score::int || ' - ' || RL.visiting_team_score::int as score,
     RL.winning_team,
     ABS(AR.elo_change) AS elo_change_num1
 FROM src_ncaaf_latest_results RL
